requirements.txt
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
Flask==3.0.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.1.5
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.3
mysql==0.0.3
mysql-connector-python==8.2.0
mysqlclient==2.2.1
protobuf==4.21.12
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.0
urllib3==2.1.0
Werkzeug==3.0.1
